bind to all addresses by default on Linux only
We expect the shim to attempt a connection to the container's gateway
address.  The uplift proxy must be listening on that address.  This is
quite a bit simpler than attempting to probe and bind to only the docker
bridge interface, so let's try the simple approach first.  I expect most
Linux distros ship with sane packet filters so this is probably fine. ;)

Does not apply to macOS and Docker Desktop.  There, a bind on the
loopback interface is sufficient.  IIRC, the macOS security nanny state
throws a spack if any process attempts a bind to all addresses, which is
a UX niggle I would rather avoid.